AMQCPP-756: Changes for C++17 compliance

+++ activemq-cpp/src/main/decaf/internal/util/zip/inflate.c
@@ -702,6 +702,7 @@ int flush;
             if (state->flags & 0x0200) CRC2(state->check, hold);
             INITBITS();
             state->mode = EXLEN;
+            /* falls through */
         case EXLEN:
             if (state->flags & 0x0400) {
                 NEEDBITS(16);
@@ -714,6 +715,7 @@ int flush;
             else if (state->head != Z_NULL)
                 state->head->extra = Z_NULL;
             state->mode = EXTRA;
+            /* falls through */
         case EXTRA:
             if (state->flags & 0x0400) {
                 copy = state->length;
@@ -736,6 +738,7 @@ int flush;
             }
             state->length = 0;
             state->mode = NAME;
+            /* falls through */
         case NAME:
             if (state->flags & 0x0800) {
                 if (have == 0) goto inf_leave;
@@ -757,6 +760,7 @@ int flush;
                 state->head->name = Z_NULL;
             state->length = 0;
             state->mode = COMMENT;
+            /* falls through */
         case COMMENT:
             if (state->flags & 0x1000) {
                 if (have == 0) goto inf_leave;
@@ -777,6 +781,7 @@ int flush;
             else if (state->head != Z_NULL)
                 state->head->comment = Z_NULL;
             state->mode = HCRC;
+            /* falls through */
         case HCRC:
             if (state->flags & 0x0200) {
                 NEEDBITS(16);
@@ -800,6 +805,7 @@ int flush;
             strm->adler = state->check = REVERSE(hold);
             INITBITS();
             state->mode = DICT;
+            /* falls through */
         case DICT:
             if (state->havedict == 0) {
                 RESTORE();
@@ -807,8 +813,10 @@ int flush;
             }
             strm->adler = state->check = adler32(0L, Z_NULL, 0);
             state->mode = TYPE;
+            /* falls through */
         case TYPE:
             if (flush == Z_BLOCK || flush == Z_TREES) goto inf_leave;
+            /* falls through */
         case TYPEDO:
             if (state->last) {
                 BYTEBITS();
@@ -859,8 +867,10 @@ int flush;
             INITBITS();
             state->mode = COPY_;
             if (flush == Z_TREES) goto inf_leave;
+            /* falls through */
         case COPY_:
             state->mode = COPY;
+            /* falls through */
         case COPY:
             copy = state->length;
             if (copy) {
@@ -1001,8 +1011,10 @@ int flush;
             Tracev((stderr, "inflate:       codes ok\n"));
             state->mode = LEN_;
             if (flush == Z_TREES) goto inf_leave;
+            /* falls through */
         case LEN_:
             state->mode = LEN;
+            /* falls through */
         case LEN:
             if (have >= 6 && left >= 258) {
                 RESTORE();
@@ -1052,6 +1064,7 @@ int flush;
             }
             state->extra = (unsigned)(here.op) & 15;
             state->mode = LENEXT;
+            /* falls through */
         case LENEXT:
             if (state->extra) {
                 NEEDBITS(state->extra);
@@ -1062,6 +1075,7 @@ int flush;
             Tracevv((stderr, "inflate:         length %u\n", state->length));
             state->was = state->length;
             state->mode = DIST;
+            /* falls through */
         case DIST:
             for (;;) {
                 here = state->distcode[BITS(state->distbits)];
@@ -1089,6 +1103,7 @@ int flush;
             state->offset = (unsigned)here.val;
             state->extra = (unsigned)(here.op) & 15;
             state->mode = DISTEXT;
+            /* falls through */
         case DISTEXT:
             if (state->extra) {
                 NEEDBITS(state->extra);
@@ -1105,6 +1120,7 @@ int flush;
 #endif
             Tracevv((stderr, "inflate:         distance %u\n", state->offset));
             state->mode = MATCH;
+            /* falls through */
         case MATCH:
             if (left == 0) goto inf_leave;
             copy = out - left;
@@ -1180,6 +1196,7 @@ int flush;
             }
 #ifdef GUNZIP
             state->mode = LENGTH;
+            /* falls through */
         case LENGTH:
             if (state->wrap && state->flags) {
                 NEEDBITS(32);
@@ -1193,6 +1210,7 @@ int flush;
             }
 #endif
             state->mode = DONE;
+            /* falls through */
         case DONE:
             ret = Z_STREAM_END;
             goto inf_leave;
@@ -1472,7 +1490,7 @@ z_streamp strm;
 {
     struct inflate_state FAR *state;
 
-    if (strm == Z_NULL || strm->state == Z_NULL) return -1L << 16;
+    if (strm == Z_NULL || strm->state == Z_NULL) return ((unsigned long)-1L) << 16;
     state = (struct inflate_state FAR *)strm->state;
     return ((long)(state->back) << 16) +
         (state->mode == COPY ? state->length :
